Pulling functions out of hyperdrive to hyperdrive agent
